refactor: moved some mocks into a file
