+++ test/batching-on-both-gateway-and-services.js
@@ -5,6 +5,7 @@ const Fastify = require('fastify')
 const GQL = require('mercurius')
 const plugin = require('../index')
 const { buildFederationSchema } = require('@mercuriusjs/federation')
+const { users, posts } = require('./utils/mocks')
 
 async function createTestService (t, schema, resolvers = {}) {
   const service = Fastify()
@@ -15,44 +16,6 @@ async function createTestService (t, schema, resolvers = {}) {
   })
   await service.listen({ port: 0 })
   return [service, service.server.address().port]
-}
-
-const users = {
-  u1: {
-    id: 'u1',
-    name: 'John'
-  },
-  u2: {
-    id: 'u2',
-    name: 'Jane'
-  }
-}
-
-const posts = {
-  p1: {
-    pid: 'p1',
-    title: 'Post 1',
-    content: 'Content 1',
-    authorId: 'u1'
-  },
-  p2: {
-    pid: 'p2',
-    title: 'Post 2',
-    content: 'Content 2',
-    authorId: 'u2'
-  },
-  p3: {
-    pid: 'p3',
-    title: 'Post 3',
-    content: 'Content 3',
-    authorId: 'u1'
-  },
-  p4: {
-    pid: 'p4',
-    title: 'Post 4',
-    content: 'Content 4',
-    authorId: 'u1'
-  }
 }
 
 async function createTestGatewayServer (t) {
