+++ test/collect-extensions.js
@@ -7,6 +7,7 @@ const GQL = require('mercurius')
 const plugin = require('../index')
 const { buildFederationSchema } = require('@mercuriusjs/federation')
 const immediate = promisify(setImmediate)
+const { users, customers, posts } = require('./utils/mocks')
 
 async function createTestService (t, schema, resolvers, serviceName) {
   const service = Fastify()
@@ -26,55 +27,6 @@ async function createTestService (t, schema, resolvers, serviceName) {
 
   await service.listen({ port: 0 })
   return [service, service.server.address().port]
-}
-
-const users = {
-  u1: {
-    id: 'u1',
-    name: 'John'
-  },
-  u2: {
-    id: 'u2',
-    name: 'Jane'
-  }
-}
-
-const customers = {
-  c1: {
-    id: 'u1',
-    name: 'John'
-  },
-  c2: {
-    id: 'u2',
-    name: 'Jane'
-  }
-}
-
-const posts = {
-  p1: {
-    pid: 'p1',
-    title: 'Post 1',
-    content: 'Content 1',
-    authorId: 'u1'
-  },
-  p2: {
-    pid: 'p2',
-    title: 'Post 2',
-    content: 'Content 2',
-    authorId: 'u2'
-  },
-  p3: {
-    pid: 'p3',
-    title: 'Post 3',
-    content: 'Content 3',
-    authorId: 'u1'
-  },
-  p4: {
-    pid: 'p4',
-    title: 'Post 4',
-    content: 'Content 4',
-    authorId: 'u1'
-  }
 }
 
 const query = `
