+++ test/retry-failed-services.js
@@ -7,6 +7,7 @@ const GQL = require('mercurius')
 const FakeTimers = require('@sinonjs/fake-timers')
 const plugin = require('../index')
 const { buildFederationSchema } = require('@mercuriusjs/federation')
+const { users, posts } = require('./utils/mocks')
 
 async function createTestService (port, schema, resolvers = {}) {
   const service = Fastify()
@@ -16,44 +17,6 @@ async function createTestService (port, schema, resolvers = {}) {
   })
   await service.listen({ port })
   return service
-}
-
-const users = {
-  u1: {
-    id: 'u1',
-    name: 'John'
-  },
-  u2: {
-    id: 'u2',
-    name: 'Jane'
-  }
-}
-
-const posts = {
-  p1: {
-    pid: 'p1',
-    title: 'Post 1',
-    content: 'Content 1',
-    authorId: 'u1'
-  },
-  p2: {
-    pid: 'p2',
-    title: 'Post 2',
-    content: 'Content 2',
-    authorId: 'u2'
-  },
-  p3: {
-    pid: 'p3',
-    title: 'Post 3',
-    content: 'Content 3',
-    authorId: 'u1'
-  },
-  p4: {
-    pid: 'p4',
-    title: 'Post 4',
-    content: 'Content 4',
-    authorId: 'u1'
-  }
 }
 
 const userService = {
